refactor(vscode): split the extension-host suite under the source budget
`editors/vscode/test/suite/extension-host.cjs` was 745 lines, more than twice
the 350-line source budget, and #4444 grew it 743 -> 745 while the budget gate
was red. The gate only compares against a base ref on pull requests, so main
never re-checked it and every branch cut before #4444 now inherits the growth:
#4452 fails `test-scripts` on a file it does not touch.

Split by what each part is rather than to hit a number:

- `extension-host-fixtures.cjs` — the published contribution surface (command
  ids, the `vize.<key>.enable` switches, the three initialization profiles)
- `extension-host-support.cjs` — the fake-server plumbing every suite shares:
  the log-polling waits, the configuration resets, the log-entry assertions
- `extension-host-grammar-smoke.cjs` — the static TextMate grammar scenario,
  the longest single suite and the only one that never starts a server
- `extension-host.cjs` — the entry point and the five lifecycle suites

No assertion changed. `editor-capability-smoke.cjs` already took this helper
surface as an argument, and it still does — `extension-host.cjs` now passes it
from the shared module instead of from its own file scope.

745 -> 300, and each new file is under the budget.
